+++ scw_js/auth_utils.ts
@@ -13,9 +13,9 @@ export interface BearerPayload {
  * Returns null if the header is missing, malformed, or fails JSON parsing.
  */
 export function parseBearerToken(authHeader: string | undefined): BearerPayload | null {
-  if (!authHeader) return null;
+  if (!authHeader) {return null;}
   const match = authHeader.match(/^Bearer\s+(.+)$/i);
-  if (!match) return null;
+  if (!match) {return null;}
   try {
     const decoded = JSON.parse(Buffer.from(match[1], "base64").toString("utf-8")) as BearerPayload;
     const { address, signature, message } = decoded;
@@ -49,13 +49,13 @@ export async function verifySignedMessage(
   expectedAddress: string,
 ): Promise<string | null> {
   const match = message.match(new RegExp(`^${expectedPrefix}:(\\d+)$`));
-  if (!match) return "Unauthorized";
+  if (!match) {return "Unauthorized";}
 
   const ts = parseInt(match[1], 10);
-  if (ts > 9_999_999_999) return "Unauthorized"; // guard against year >2286 / integer overflow
-  if (Math.abs(Date.now() - ts * 1000) > MAX_AUTH_AGE_MS) return "Token expired";
+  if (ts > 9_999_999_999) {return "Unauthorized";} // guard against year >2286 / integer overflow
+  if (Math.abs(Date.now() - ts * 1000) > MAX_AUTH_AGE_MS) {return "Token expired";}
 
-  if (address.toLowerCase() !== expectedAddress.toLowerCase()) return "Address mismatch";
+  if (address.toLowerCase() !== expectedAddress.toLowerCase()) {return "Address mismatch";}
 
   try {
     const isValid = await verifyMessage({
@@ -63,7 +63,7 @@ export async function verifySignedMessage(
       message,
       signature: signature as `0x${string}`,
     });
-    if (!isValid) return "Invalid signature";
+    if (!isValid) {return "Invalid signature";}
   } catch {
     return "Invalid signature";
   }
