+++ scw_js/getChain.ts
@@ -1,50 +1,34 @@
-// @ts-check
 import { sepolia, optimism, optimismSepolia } from "viem/chains";
+import type { Chain } from "viem";
 import {
   LLMv1ABI,
   getGenAiNFTMainnetNetworks,
   getGenAiNFTTestnetNetworks,
 } from "@fretchen/chain-utils";
 
-/**
- * Get environment variable in both Node.js and Vite contexts
- * @param {string} key
- * @param {string} defaultValue
- * @returns {string}
- */
-function getEnvironmentVariable(key, defaultValue) {
+function getEnvironmentVariable(key: string, defaultValue: string): string {
   try {
-    // Try Vite environment (browser/build context)
-    if (typeof import.meta !== "undefined" && import.meta.env) {
-      return import.meta.env[key] || defaultValue;
+    if (typeof import.meta !== "undefined" && (import.meta as unknown as { env?: Record<string, string> }).env) {
+      return (import.meta as unknown as { env: Record<string, string> }).env[key] ?? defaultValue;
     }
   } catch {
-    // Fallback to Node.js environment
+    // fallback
   }
-
   try {
-    // Try Node.js environment
     if (typeof process !== "undefined" && process.env) {
-      return process.env[key] || defaultValue;
+      return process.env[key] ?? defaultValue;
     }
   } catch {
-    // Fallback to default
+    // fallback
   }
-
   return defaultValue;
 }
 
 /**
- * Gibt das entsprechende Chain-Objekt basierend auf der CHAIN-Umgebungsvariable zurück.
- *
  * @deprecated TODO: Migrate to CAIP-2 pattern when LLMv1 gets multi-chain support.
- * Currently used by:
- * - llm_service.js: checkWalletBalance() (line ~454)
- * - llm_service.js: processMerkleTree() (line ~572)
- *
- * @returns {import("viem/chains").Chain}
+ * Currently used by llm_service.ts checkWalletBalance() and processMerkleTree().
  */
-export function getChain() {
+export function getChain(): Chain {
   const chainName = getEnvironmentVariable("PUBLIC_ENV__CHAIN_NAME", "optimism");
   switch (chainName) {
     case "sepolia":
@@ -58,10 +42,7 @@ export function getChain() {
   }
 }
 
-/** returns the config of the LLMv1 contract
- * @returns {{ address: `0x${string}`, abi: any }}
- */
-export function getLLMv1ContractConfig() {
+export function getLLMv1ContractConfig(): { address: `0x${string}`; abi: typeof LLMv1ABI } {
   const chainName = getEnvironmentVariable("PUBLIC_ENV__CHAIN_NAME", "optimism");
   switch (chainName) {
     case "optimismSepolia":
@@ -73,37 +54,28 @@ export function getLLMv1ContractConfig() {
   }
 }
 
-/**
- * Get the expected networks for a given mode
- * Dynamically pulls from chain-utils deployment configuration
- * @param {boolean} sepoliaTest - Whether test mode is enabled
- * @returns {string[]} Array of CAIP-2 network IDs
- */
-export function getExpectedNetworks(sepoliaTest) {
+export function getExpectedNetworks(sepoliaTest: boolean): readonly string[] {
   return sepoliaTest ? getGenAiNFTTestnetNetworks() : getGenAiNFTMainnetNetworks();
 }
 
-/**
- * Validate that a client-selected network is supported for the current mode
- * @param {string|undefined} clientNetwork - Network from payment payload
- * @param {boolean} sepoliaTest - Whether test mode is enabled
- * @returns {{ valid: boolean, reason?: string, expected?: string[], received?: string }}
- */
-export function validatePaymentNetwork(clientNetwork, sepoliaTest = false) {
+export type NetworkValidationResult =
+  | { valid: true }
+  | { valid: false; reason: string; expected?: readonly string[]; received?: string };
+
+export function validatePaymentNetwork(
+  clientNetwork: string | undefined,
+  sepoliaTest = false,
+): NetworkValidationResult {
   if (!clientNetwork) {
     return { valid: false, reason: "missing_network" };
   }
 
-  // Get expected networks for the current mode
   const expectedNetworks = getExpectedNetworks(sepoliaTest);
 
-  // Check if the client network matches the expected mode
   if (!expectedNetworks.includes(clientNetwork)) {
-    // Determine if it's an unsupported network or wrong mode
     const allNetworks = [...getExpectedNetworks(false), ...getExpectedNetworks(true)];
 
     if (allNetworks.includes(clientNetwork)) {
-      // Network exists but wrong mode
       return {
         valid: false,
         reason: sepoliaTest ? "invalid_network_for_test_mode" : "invalid_network_for_production",
@@ -112,7 +84,6 @@ export function validatePaymentNetwork(clientNetwork, sepoliaTest = false) {
       };
     }
 
-    // Completely unsupported network
     return {
       valid: false,
       reason: "unsupported_network",
