+++ scw_js/growth_api.ts
@@ -9,8 +9,8 @@ import {
   verifyOwner,
   NotFoundError,
   AuthError,
+  type ContentQueue,
 } from "./growth_service.js";
-import type { ContentQueue } from "./growth_service.js";
 import { parseBearerToken } from "./auth_utils.js";
 
 const logger = pino({ level: process.env.LOG_LEVEL || "info" });
@@ -44,10 +44,10 @@ function filterByStatus(queue: ContentQueue, status: string | undefined): unknow
 }
 
 function parseJsonBody(raw: unknown): Record<string, unknown> | null {
-  if (raw == null || raw === "") return null;
+  if (raw === null || raw === undefined || raw === "") {return null;}
   try {
     const parsed = typeof raw === "string" ? JSON.parse(raw) : raw;
-    if (typeof parsed !== "object" || Array.isArray(parsed) || parsed === null) return null;
+    if (typeof parsed !== "object" || Array.isArray(parsed) || parsed === null) {return null;}
     return parsed as Record<string, unknown>;
   } catch {
     return null;
