+++ scw_js/image_service.ts
@@ -1,10 +1,5 @@
-/**
- * Shared module for image generation and S3 upload
- */
-// Only load dotenv if .env file exists and not in CI environment
 if (process.env.NODE_ENV === "test" && !process.env.CI) {
   try {
-    // Since we're in an ESM context, we need to use dynamic import
     await import("dotenv").then((dotenv) => {
       dotenv.config();
       console.log("Environment variables loaded from .env");
@@ -17,8 +12,15 @@ if (process.env.NODE_ENV === "test" && !process.env.CI) {
 import { S3Client, PutObjectCommand } from "@aws-sdk/client-s3";
 import { randomBytes } from "crypto";
 
-// Provider configurations
-const PROVIDER_CONFIGS = {
+type Provider = "ionos" | "bfl";
+
+interface ProviderConfig {
+  endpoint: string;
+  model: string;
+  tokenEnvVar: string;
+}
+
+const PROVIDER_CONFIGS: Record<Provider, ProviderConfig> = {
   ionos: {
     endpoint: "https://openai.inference.de-txl.ionos.com/v1/images/generations",
     model: "black-forest-labs/FLUX.1-schnell",
@@ -34,37 +36,24 @@ const PROVIDER_CONFIGS = {
 export const JSON_BASE_PATH = "https://my-imagestore.s3.nl-ams.scw.cloud/";
 const BUCKET_NAME = "my-imagestore";
 
-/**
- * Generates a random string for filenames
- */
-function getRandomString(length = 6) {
+function getRandomString(length = 6): string {
   return randomBytes(length).toString("hex");
 }
 
-/**
- * Converts a Base64 string to a Buffer for binary upload
- * @param {string} base64String - The Base64 string to convert
- * @returns {Buffer} - Binary buffer
- */
-function base64ToBuffer(base64String) {
-  // Remove potential data URL prefix
+function base64ToBuffer(base64String: string): Buffer {
   const base64Data = base64String.replace(/^data:image\/\w+;base64,/, "");
   return Buffer.from(base64Data, "base64");
 }
 
-/**
- * Uploads any data object to S3.
- * @param {Object|string|Buffer} data - The data to upload
- * @param {string} fileName - The filename to save the data as
- * @param {string} contentType - MIME type of the content
- * @returns {Promise<string>} - Path to the uploaded file
- */
-export async function uploadToS3(data, fileName, contentType = "application/json") {
+export async function uploadToS3(
+  data: Buffer | object | string,
+  fileName: string,
+  contentType = "application/json",
+): Promise<string> {
   const accessKey = process.env.SCW_ACCESS_KEY;
   const secretKey = process.env.SCW_SECRET_KEY;
 
-  // Prepare data for upload based on type
-  let dataToUpload;
+  let dataToUpload: Buffer | string;
   if (Buffer.isBuffer(data)) {
     dataToUpload = data;
   } else if (typeof data === "object") {
@@ -73,26 +62,23 @@ export async function uploadToS3(data, fileName, contentType = "application/json
     dataToUpload = data;
   }
 
-  // Create the S3 client
   const s3Client = new S3Client({
     region: "nl-ams",
     endpoint: "https://s3.nl-ams.scw.cloud",
     credentials: {
-      accessKeyId: accessKey,
-      secretAccessKey: secretKey,
+      accessKeyId: accessKey!,
+      secretAccessKey: secretKey!,
     },
   });
 
-  // Configure the upload
   const params = {
     Bucket: BUCKET_NAME,
     Key: fileName,
     Body: dataToUpload,
     ContentType: contentType,
-    ACL: "public-read",
+    ACL: "public-read" as const,
   };
 
-  // Perform the upload
   try {
     await s3Client.send(new PutObjectCommand(params));
     console.log(`Successfully uploaded ${fileName}`);
@@ -103,13 +89,7 @@ export async function uploadToS3(data, fileName, contentType = "application/json
   }
 }
 
-/**
- * Generates an image using IONOS API
- * @param {string} prompt - The prompt for image generation
- * @param {string} size - Image size
- * @returns {Promise<string>} - Base64 encoded image
- */
-async function generateImageIONOS(prompt, size) {
+async function generateImageIONOS(prompt: string, size: string): Promise<string> {
   const config = PROVIDER_CONFIGS.ionos;
   const apiToken = process.env[config.tokenEnvVar];
 
@@ -119,41 +99,29 @@ async function generateImageIONOS(prompt, size) {
     );
   }
 
-  const headers = {
-    Authorization: `Bearer ${apiToken}`,
-    "Content-Type": "application/json",
-  };
-
-  const body = {
-    model: config.model,
-    prompt,
-    size,
-  };
+  const body = { model: config.model, prompt, size };
 
   console.log("Sending IONOS image generation request...");
   const response = await fetch(config.endpoint, {
     method: "POST",
-    headers,
+    headers: { Authorization: `Bearer ${apiToken}`, "Content-Type": "application/json" },
     body: JSON.stringify(body),
-    timeout: 60000,
   });
 
   if (!response.ok) {
-    console.error(`IONOS API Error: ${response.status} ${response.statusText}`);
     throw new Error(`Could not reach IONOS: ${response.status} ${response.statusText}`);
   }
 
-  const responseData = await response.json();
-  return responseData.data[0].b64_json;
+  const responseData = await response.json() as { data: Array<{ b64_json: string }> };
+  return responseData.data[0]!.b64_json;
 }
 
-/**
- * Generates an image using BFL API with proper polling
- * @param {string} prompt - The prompt for image generation
- * @param {string} size - Image size
- * @returns {Promise<string>} - Base64 encoded image
- */
-async function generateImageBFL(prompt, size, mode = "generate", referenceImageBase64 = null) {
+async function generateImageBFL(
+  prompt: string,
+  size: string,
+  mode = "generate",
+  referenceImageBase64: string | null = null,
+): Promise<string> {
   const config = PROVIDER_CONFIGS.bfl;
   const apiToken = process.env[config.tokenEnvVar];
 
@@ -165,48 +133,43 @@ async function generateImageBFL(prompt, size, mode = "generate", referenceImageB
 
   console.log(`Sending BFL image generation request in ${mode} mode...`);
 
-  // Prepare request body based on mode
-  const requestBody = {
+  const requestBody: Record<string, unknown> = {
     prompt,
-    aspect_ratio: size === "1792x1024" ? "16:9" : "1:1", // Use aspect_ratio instead of width/height
-    output_format: "jpeg", // Ensure consistent JPEG format
+    aspect_ratio: size === "1792x1024" ? "16:9" : "1:1",
+    output_format: "jpeg",
   };
 
-  // Add reference image for edit mode
   if (mode === "edit" && referenceImageBase64) {
-    requestBody.input_image = referenceImageBase64;
+    requestBody["input_image"] = referenceImageBase64;
     console.log("Reference image added for editing");
   }
 
-  // Step 1: Start image generation
   const response = await fetch(config.endpoint, {
     method: "POST",
     headers: {
       accept: "application/json",
-      "x-key": apiToken, // BFL API requires authentication via the non-standard 'x-key' header instead of the standard 'Authorization' header. See BFL API docs for details.
+      // BFL API requires authentication via non-standard 'x-key' header. See BFL API docs.
+      "x-key": apiToken,
       "Content-Type": "application/json",
     },
     body: JSON.stringify(requestBody),
   });
 
   if (!response.ok) {
-    console.error(`BFL API Error: ${response.status} ${response.statusText}`);
     throw new Error(`Could not reach BFL: ${response.status} ${response.statusText}`);
   }
 
-  const initData = await response.json();
+  const initData = await response.json() as { id: string; polling_url: string };
   const { id: requestId, polling_url } = initData;
 
   console.log(`BFL request started with ID: ${requestId}`);
 
-  // Step 2: Poll for completion
-  const maxAttempts = 60; // 5 minutes max (60 * 5 seconds)
-  const pollInterval = 5000; // 5 seconds
+  const maxAttempts = 60;
+  const pollInterval = 5000;
 
   for (let attempt = 0; attempt < maxAttempts; attempt++) {
     console.log(`Polling attempt ${attempt + 1}/${maxAttempts}...`);
 
-    // Wait before polling (except first attempt)
     if (attempt > 0) {
       await new Promise((resolve) => setTimeout(resolve, pollInterval));
     }
@@ -214,25 +177,22 @@ async function generateImageBFL(prompt, size, mode = "generate", referenceImageB
     try {
       const pollResponse = await fetch(polling_url, {
         method: "GET",
-        headers: {
-          accept: "application/json",
-          "x-key": apiToken,
-        },
+        headers: { accept: "application/json", "x-key": apiToken },
       });
 
       if (!pollResponse.ok) {
         console.warn(`Poll request failed: ${pollResponse.status}`);
-        continue; // Retry on next attempt
+        continue;
       }
 
-      const pollData = await pollResponse.json();
+      const pollData = await pollResponse.json() as {
+        status: string;
+        result?: { sample: string };
+      };
       console.log(`Poll status: ${pollData.status}`);
 
       if (pollData.status === "Ready") {
-        console.log("BFL image generation completed!");
-        const imageUrl = pollData.result.sample;
-
-        // Step 3: Download the actual image
+        const imageUrl = pollData.result!.sample;
         console.log("Downloading image from:", imageUrl);
         const imageResponse = await fetch(imageUrl);
 
@@ -241,19 +201,12 @@ async function generateImageBFL(prompt, size, mode = "generate", referenceImageB
         }
 
         const imageBuffer = await imageResponse.arrayBuffer();
-        const base64Image = Buffer.from(imageBuffer).toString("base64");
-
-        console.log("BFL image successfully downloaded and encoded");
-        return base64Image;
+        return Buffer.from(imageBuffer).toString("base64");
       } else if (pollData.status === "Error" || pollData.status === "Failed") {
-        console.error("BFL generation failed:", pollData);
         throw new Error(`BFL generation failed: ${JSON.stringify(pollData)}`);
       }
-
-      // Status is still "processing" or similar, continue polling
     } catch (error) {
-      console.warn(`Polling error (attempt ${attempt + 1}):`, error.message);
-      // Continue to next attempt
+      console.warn(`Polling error (attempt ${attempt + 1}):`, (error as Error).message);
     }
   }
 
@@ -262,57 +215,38 @@ async function generateImageBFL(prompt, size, mode = "generate", referenceImageB
   );
 }
 
-/**
- * Generates an image from the specified provider
- * @param {string} prompt - The prompt for image generation
- * @param {string} provider - The provider to use ('ionos' or 'bfl')
- * @param {string} size - Image size
- * @returns {Promise<string>} - Base64 encoded image
- */
 async function generateImageFromProvider(
-  prompt,
-  provider,
-  size,
+  prompt: string,
+  provider: Provider,
+  size: string,
   mode = "generate",
-  referenceImageBase64 = null,
-) {
+  referenceImageBase64: string | null = null,
+): Promise<string> {
   switch (provider) {
     case "ionos":
       return generateImageIONOS(prompt, size);
     case "bfl":
       return generateImageBFL(prompt, size, mode, referenceImageBase64);
-    default:
-      throw new Error(`Unknown provider: ${provider}`);
   }
 }
 
-/**
- * Generates an image based on the prompt and uploads it to S3 along with ERC-721 metadata
- * @param {string} prompt - The prompt for image generation
- * @param {string|number} tokenId - The NFT token ID to include in metadata
- * @param {string} provider - The image generation provider ('ionos' or 'bfl')
- * @param {string} size - Image size, either "1024x1024" or "1792x1024"
- * @returns {Promise<string>} - Path to the generated metadata file
- */
 export async function generateAndUploadImage(
-  prompt,
-  tokenId = "unknown",
-  provider,
+  prompt: string,
+  tokenId: string | number = "unknown",
+  provider: Provider,
   size = "1024x1024",
   mode = "generate",
-  referenceImageBase64 = null,
-) {
+  referenceImageBase64: string | null = null,
+): Promise<string> {
   if (!prompt) {
     throw new Error("No prompt provided.");
   }
 
-  // Validate size parameter
   const validSizes = ["1024x1024", "1792x1024"];
   if (!validSizes.includes(size)) {
     throw new Error(`Invalid size parameter. Must be one of: ${validSizes.join(", ")}`);
   }
 
-  // Generate image using the specified provider
   const imageBase64 = await generateImageFromProvider(
     prompt,
     provider,
@@ -322,44 +256,24 @@ export async function generateAndUploadImage(
   );
   console.log("Image received from", provider, "in", mode, "mode");
 
-  // Upload the image as JPEG in the images subfolder
   const imageFileName = `images/image_${tokenId}_${getRandomString()}.jpg`;
   const imageBuffer = base64ToBuffer(imageBase64);
   const imageUrl = await uploadToS3(imageBuffer, imageFileName, "image/jpeg");
 
-  // Create and upload ERC-721 compliant metadata in the metadata subfolder
   const metadataFileName = `metadata/metadata_${tokenId}_${getRandomString()}.json`;
-
-  // Create metadata following ERC-721 standard
   const metadata = {
     name: `AI Generated Art #${tokenId}`,
     description: `AI generated artwork based on the prompt: "${prompt}"`,
-    image: imageUrl, // Reference to the JPEG image
+    image: imageUrl,
     attributes: [
-      {
-        trait_type: "Prompt",
-        value: prompt,
-      },
-      {
-        trait_type: "Model",
-        value: PROVIDER_CONFIGS[provider].model,
-      },
-      {
-        trait_type: "Image Size",
-        value: size,
-      },
-      {
-        trait_type: "Creation Date",
-        value: new Date().toISOString(),
-      },
+      { trait_type: "Prompt", value: prompt },
+      { trait_type: "Model", value: PROVIDER_CONFIGS[provider].model },
+      { trait_type: "Image Size", value: size },
+      { trait_type: "Creation Date", value: new Date().toISOString() },
     ],
   };
 
-  // Upload the metadata
   const metadataUrl = await uploadToS3(metadata, metadataFileName);
-
   console.log(`Image and metadata uploaded successfully for token ${tokenId}`);
-
-  // Return the metadata URL (which contains a reference to the image)
   return metadataUrl;
 }
