+++ scw_js/package.json
@@ -13,8 +13,8 @@
     "lint:fix": "eslint . --fix",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "check": "npm run lint && npm run format:check && npm run test:coverage",
-    "dev:x402": "NODE_ENV=test node genimg_x402_token.js",
+    "check": "tsc --noEmit && npm run lint && npm run format:check && npm run test:coverage",
+    "dev:x402": "NODE_ENV=test npx tsx genimg_x402_token.ts",
     "dev:growth": "NODE_ENV=test npx tsx growth_api.ts"
   },
   "keywords": [],
