+++ scw_js/test/image_service.test.ts
