+++ scw_js/test/llm_service.test.ts
