+++ scw_js/test/nft_abi.test.ts
